The Skills Update
So, I finally feel satisfied with the Skills chapter, and it's uploaded (w00t!). I hate to say this, but adjust characters accoringly. I will have to update Advancement.md and I think CCSummary.md, and I'm going to add in a Dice-For-Success trading mechanic to help combat potentially silly numbers of dice. But that's for next update.

Altered all Talent names to be adjectives so that they can be used to describe your character (what fun!), gave sample skills an instructions on the parameters for all skills, re-worked the Skill Point progression to something sane and easy to read, reworked Foci to be more comprehensible and (hopefully) balanced. Added Dodge and Evade as special skills.