Effort and Energy
Updated Effort and Energy, started on Engagement.md. Also made some modifications to personal "hidden" resources in the T00nz folder - tweaks to the Scion2e toonsheet, and the start of a log for an upcoming game. Not sure if it's okay to use the repo this way, but here we are.