Engagements DONE
Alright, I updated Energy.md, and actually finished Engagement.md. I could really use someone to read through it an make sure it makes sense. Otherwise, it just needs a PLaytest (like everything else, yay).

There's a "hidden" file with an older version of Engagements (well, part of it, anyway) that seemed fun while I was writing it but was a lot more complicated than it needed to be. It's _alt.complex.engagement.md if you're interested. I need to make a "Historical" folder for all this old-version stuff.